moving countdown timer into proper location
